Integrated user to db tables and added logout

+++ modify_db.py
@@ -2,6 +2,7 @@ import sqlite3
 import streamlit as st
 import pandas as pd
 
+
 #####################################################################
 #                Methods for storing and retrieving user data
 #####################################################################
@@ -20,6 +21,7 @@ def create_user_table():
     conn.commit()
     conn.close()
 
+
 def add_user(username, password):
     """
     Registers a new user into the database.
@@ -35,6 +37,7 @@ def add_user(username, password):
         st.error(f"Username '{username}' is already taken. Please choose a different username.")
     conn.close()
 
+
 def check_user_credentials(username, password):
     """
     Returns True if username+password exist in the database, else False.
@@ -44,30 +47,40 @@ def check_user_credentials(username, password):
     cursor.execute("SELECT * FROM users WHERE username=? AND password=?", (username, password))
     row = cursor.fetchone()
     conn.close()
-    return row is not None
-
+    # return row is not None
+    return row
 
 #####################################################################
 #       Methods to add, retrieve, and remove calendar events
 #####################################################################
-def add_event(title, date, date_start, date_end):
+def create_event_db():
     # Connect to SQLite database (creates file if not exists)
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
-    # conn.execute("DROP TABLE events")
     # Create a table if not exists
     cursor.execute("""
         CREATE TABLE IF NOT EXISTS events (
             id INTEGER PRIMARY KEY AUTOINCREMENT,
+            user_id INTEGER,
             title TEXT,
             date TEXT,
             datetime_start TEXT,
-            datetime_end TEXT
+            datetime_end TEXT,
+            FOREIGN KEY(user_id) REFERENCES users(id)
         )
     """)
     conn.commit()
-    cursor.execute("INSERT INTO events (title, date, datetime_start, datetime_end) VALUES (?, ?, ?, ?)",
-                   (title, date, date_start, date_end))
+
+    # Close the connection
+    conn.close()
+
+
+def add_event(user_id, title, date, date_start, date_end):
+    # Connect to SQLite database (creates file if not exists)
+    conn = sqlite3.connect("Calendar.db")
+    cursor = conn.cursor()
+    cursor.execute("INSERT INTO events (user_id, title, date, datetime_start, datetime_end) VALUES (?, ?, ?, ?, ?)",
+                   (user_id, title, date, date_start, date_end))
     conn.commit()
     st.success(f"Event '{title}' added!")
 
@@ -75,14 +88,14 @@ def add_event(title, date, date_start, date_end):
     conn.close()
 
 
-def check_time_exists(date_start, date_end):
+def check_time_exists(user_id, date_start, date_end):
     try:
         conn = sqlite3.connect("Calendar.db")
         cursor = conn.cursor()
 
         cursor.execute(
-            "SELECT COUNT(*) FROM events WHERE datetime_start = ? OR datetime_end = ?",
-            (date_start, date_end)
+            "SELECT COUNT(*) FROM events WHERE (datetime_start = ? OR datetime_end = ?) AND user_id=?",
+            (date_start, date_end, user_id)
         )
         result = cursor.fetchone()[0]
         conn.close()
@@ -95,23 +108,12 @@ def check_time_exists(date_start, date_end):
         return False
 
 
-def show_events():
+def show_events(user_id):
     # Connect to SQLite database
     conn = sqlite3.connect("Calendar.db")
-    cursor = conn.cursor()
     try:
-        cursor.execute("""
-                CREATE TABLE IF NOT EXISTS events (
-                    id INTEGER PRIMARY KEY AUTOINCREMENT,
-                    title TEXT,
-                    date TEXT,
-                    datetime_start TEXT,
-                    datetime_end TEXT
-                )
-            """)
-        conn.commit()
         # Create a dataframe of calendar events
-        df = pd.read_sql_query("SELECT * FROM events", conn)
+        df = pd.read_sql_query(f"SELECT * FROM events WHERE user_id={user_id}", conn)
         return df
     except sqlite3.IntegrityError:
         print("No events available")
@@ -121,25 +123,16 @@ def show_events():
     return
 
 
-def remove_event(title, date, time):
+def remove_event(user_id, title, date, time):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     try:
-        cursor.execute("""
-                        CREATE TABLE IF NOT EXISTS events (
-                            id INTEGER PRIMARY KEY AUTOINCREMENT,
-                            title TEXT,
-                            date TEXT,
-                            datetime_start TEXT,
-                            datetime_end TEXT
-                        )
-                    """)
         if title and date and time:
-            cursor.execute("DELETE FROM events WHERE title = ? AND date = ? AND datetime_start = ?",
-                           (title, date, time))
+            cursor.execute("DELETE FROM events WHERE user_id = ? AND title = ? AND date = ? AND datetime_start = ?",
+                           (user_id, title, date, time))
         elif title and date:
-            cursor.execute("DELETE FROM events WHERE title = ? AND date = ?",
-                           (title, date))
+            cursor.execute("DELETE FROM events WHERE user_id=? AND title = ? AND date = ?",
+                           (user_id, title, date))
             st.success(f"{title} removed")
         conn.commit()
         st.success(f"The following event has been removed: '{title}' ")
@@ -153,20 +146,28 @@ def remove_event(title, date, time):
 #####################################################################
 #       Methods to add and remove homework assignments due          #
 #####################################################################
-def add_hw(title, due_date):
+def create_assignment_db():
+    conn = sqlite3.connect("Calendar.db")
+    cursor = conn.cursor()
+    cursor.execute("""
+                        CREATE TABLE IF NOT EXISTS assignments (
+                            id INTEGER PRIMARY KEY AUTOINCREMENT,
+                            user_id INTEGER,
+                            title TEXT UNIQUE,
+                            date_due TEXT,
+                            FOREIGN KEY(user_id) REFERENCES users(id)
+                        )
+                    """)
+    conn.commit()
+    conn.close()
+
+
+def add_hw(user_id, title, due_date):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     try:
-        cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS assignments (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        title TEXT UNIQUE,
-                        date_due TEXT
-                    )
-                """)
-        conn.commit()
-        cursor.execute("INSERT INTO assignments (title, date_due) VALUES (?, ?)",
-                       (title, due_date))
+        cursor.execute("INSERT INTO assignments (user_id, title, date_due) VALUES (?, ?, ?)",
+                       (user_id, title, due_date))
         conn.commit()
         st.success("Your assignment due date has been scheduled.")
     except sqlite3.IntegrityError:
@@ -175,22 +176,14 @@ def add_hw(title, due_date):
     conn.close()
 
 
-def get_hw():
+def get_hw(user_id):
     # Connect to SQLite database
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     # conn.execute("DROP TABLE assignments")
     try:
-        cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS assignments (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        title TEXT UNIQUE,
-                        date_due TEXT
-                    )
-                """)
-        conn.commit()
         # Create a dataframe of calendar events
-        df = pd.read_sql_query("SELECT * FROM assignments", conn)
+        df = pd.read_sql_query(f"SELECT * FROM assignments WHERE user_id={user_id}", conn)
         return df
     except sqlite3.IntegrityError:
         print("No assignment are currently due.")
@@ -200,20 +193,12 @@ def get_hw():
     return
 
 
-def remove_hw(title):
+def remove_hw(user_id, title):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     try:
-        cursor.execute("""
-                        CREATE TABLE IF NOT EXISTS assignments (
-                            id INTEGER PRIMARY KEY AUTOINCREMENT,
-                            title TEXT UNIQUE,
-                            date_due TEXT
-                        )
-                    """)
         for hw in title:
-            print(hw)
-            cursor.execute("DELETE FROM assignments WHERE title = (?) ", (hw,))
+            cursor.execute("DELETE FROM assignments WHERE user_id = ? AND title = (?) ", (user_id, hw))
             conn.commit()
     except sqlite3.IntegrityError:
         print("That assignment does not exist.")
@@ -225,19 +210,27 @@ def remove_hw(title):
 #####################################################################
 #       Methods to add and remove tasks from the to-do list         #
 #####################################################################
-def add_todo(task):
+def create_todo_db():
+    conn = sqlite3.connect("Calendar.db")
+    cursor = conn.cursor()
+    # Create a table if not exists
+    cursor.execute("""
+                CREATE TABLE IF NOT EXISTS todo (
+                    id INTEGER PRIMARY KEY AUTOINCREMENT,
+                    user_id INTEGER,
+                    title TEXT,
+                    FOREIGN KEY(user_id) REFERENCES users(id)
+                )
+            """)
+    conn.commit()
+    conn.close()
+
+
+def add_todo(user_id, task):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     # conn.execute("DROP TABLE todo") # for testing
-    # Create a table if not exists
-    cursor.execute("""
-            CREATE TABLE IF NOT EXISTS todo (
-                id INTEGER PRIMARY KEY AUTOINCREMENT,
-                title TEXT
-            )
-        """)
-    conn.commit()
-    cursor.execute("INSERT INTO todo (title) VALUES (?)", (task,))
+    cursor.execute("INSERT INTO todo (user_id, title) VALUES (?,?)", (user_id, task))
     conn.commit()
     st.success(f"Task '{task}' added!")
 
@@ -245,17 +238,11 @@ def add_todo(task):
     conn.close()
 
 
-def remove_task(task):
+def remove_task(user_id, task):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
     try:
-        cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS todo (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        title TEXT
-                    )
-                """)
-        cursor.execute("DELETE FROM todo WHERE title = ?", (task,))
+        cursor.execute("DELETE FROM todo WHERE title = ? AND user_id=?", (task, user_id))
         conn.commit()
     except sqlite3.IntegrityError:
         print()
@@ -263,17 +250,13 @@ def remove_task(task):
     # Close the connection
     conn.close()
 
-def get_todos():
+
+def get_todos(user_id):
     conn = sqlite3.connect("Calendar.db")
     cursor = conn.cursor()
+    tasks = []  # Initialize tasks as an empty list
     try:
-        cursor.execute("""
-                    CREATE TABLE IF NOT EXISTS todo (
-                        id INTEGER PRIMARY KEY AUTOINCREMENT,
-                        title TEXT
-                    )
-                """)
-        cursor.execute("SELECT title FROM todo")
+        cursor.execute("SELECT title FROM todo WHERE user_id =?", (user_id,))
         tasks = cursor.fetchall()  # tasks is a list of tuples
     except sqlite3.Error as e:
         print(e)
